+++ spiderweb/scripts/create_mysql_index.sql
@@ -1,0 +1,4 @@
+/* If you would to increase speed on search of a callsign spot,
+   you could add the following indexes on your dxcluster database
+*/
+CREATE INDEX spiderweb_spotter ON spot (spotter);