+++ spiderweb/scripts/qso_trend.sh
@@ -1,0 +1,5 @@
+DIR=$(realpath -s $0|sed 's|\(.*\)/.*|\1|')
+echo Absolute path: ${DIR}
+cd ${DIR} 
+python3 ../lib/qso_trend.py
+