+++ spiderweb/static/js/copy_date.js
@@ -1,0 +1,1 @@
+document.getElementById('copyDate').innerHTML='2020-'.concat(new Date().getFullYear())