+++ spiderweb/test.sh
@@ -1,0 +1,7 @@
+if [ "$1" == "-b" ]; then
+	cd scripts || exit
+	./build.sh
+	cd ..
+fi
+python3 webapp.py
+