+++ spiderweb/wsgi.py
@@ -1,0 +1,6 @@
+import bjoern
+from webapp import app
+
+
+bjoern.listen(app, '0.0.0.0', 8080)
+bjoern.run()